requirements.txt
bcrypt==5.0.0
blinker==1.9.0
bson==0.5.10
click==8.3.0
dnspython==2.8.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.10.1
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
Werkzeug==3.1.3
